new example to demonstrate the backup of the RTC on stm32F1
On the stm32F1, this example saves the date/time of the current RTC.
With flag true, the BackUp Registers are cleared. Then on the next reset
or a cold reset the RTC starts the from initial date.
With Flag false, on the next (warm) reset, the calendar starts
from the (latest) saved value.

Signed-off-by: Francois Ramu <francois.ramu@st.com>