store and retrieve the current RTC into backup reg on stm32F1
The stm32F1 RTC counter registers are read/write from/to a pair
(RTC_BKP_SAVE_H and RTC_BKP_SAVE_L) of BackUp registers.
This is possible with exporting the stm32F1xx_HAL_rtc functions
RTC_WriteTimeCounter and RTC_ReadTimeCounter.
A callback is attached on the One-second interrupt to update
the bacKup value.

Signed-off-by: Francois Ramu <francois.ramu@st.com>